+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/ext/RoomDatabaseExt.kt
@@ -229,35 +229,3 @@ suspend fun RoomDatabase.resetRoom(newVersion: Int = 0) {
 suspend fun RoomDatabase.isIntegrityOk(): Boolean {
     return useReaderConnection { it.isIntegrityOk() }
 }
-
-/**
- * If the database should NOT have a migration and is a pre-populated database that should not be managed by Room... make sure Room migration is never needed.
- *
- * NOTE: this SHOULD be called BEFORE room has a chance to open the database and verify the database
- *
- * Example Usage:
- *     val driver = BundledSQLiteDriver()
- *     val connection = driver.open(mySqliteDbFileName)
- *     connection.checkAndFixRoomIdentityHash(MyDatabase.DATABASE_VERSION, MyDatabase.ROOM_DATABASE_IDENTITY_HASH)
- *     connection.close()
- *
- * OR Example using SQLiteDriverExt
- *     val driver = BundledSQLiteDriver()
- *     driver.checkAndFixRoomIdentityHash(mySqliteDbFileName, MyDatabase.DATABASE_VERSION, MyDatabase.ROOM_DATABASE_IDENTITY_HASH)
- *
- * @param expectedVersion SQLite Database version (PRAGMA user_version)
- * @param expectedIdentityHash Hash that is expected.  If the expectedIdentityHash does not match the existing identity hash (currently in the room_master_table), then just delete the table
- */
-suspend fun RoomDatabase.checkAndFixRoomIdentityHash(expectedVersion: Int, expectedIdentityHash: String) {
-    return useWriterConnection { it.checkAndFixRoomIdentityHash(expectedVersion, expectedIdentityHash) }
-}
-
-/**
- * Find the Room Identity Hash
- * Note: if you are not sure if the room_master_table exists, check first with tableExists(database, "room_master_table")
- *
- * @return identity_hash for this database OR null if it does exist
- */
-suspend fun RoomDatabase.findRoomIdentityHash(): String? {
-    return useReaderConnection { it.findRoomIdentityHash() }
-}